requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pygame==2.5.2
pillow==10.0.1
plyer==2.1.0
matplotlib==3.7.2
ultralytics==8.0.196
torch==2.0.1
torchvision==0.15.2
psycopg2-binary==2.9.7
flask==2.3.3
flask-socketio==5.3.4
requests==2.31.0
